Split availability state by responsibility
Separate app, commit, epoch-transition, lane, and persistence logic so the state orchestration remains reviewable.
